// src/types/demux.ts
// Generated by https://quicktype.io

export interface ServiceRequest {
  service: string;
  data: Buffer;
}

export interface Request {
  requestId: number;
  serviceRequest: ServiceRequest;
}

export interface ServiceRequestUpstream {
  request: Request;
}

export interface ServiceResponse {
  success: boolean;
  data: Buffer;
}

export interface Response {
  requestId: number;
  serviceRsp: ServiceResponse;
}

export interface ServiceResponseDownstream {
  response: Response;
}